requirements.txt
requests~=2.31.0
fpdf2~=2.7.4
